Add files via upload
FIXED an important error. I'd somehow left parameter `z` in there -- and in the data block, which is even crazier.  That parameter (`z`) is the latent state of the environment -- occupied or not, and so coded as 0 or 1 -- and it is standard in many of these occupancy models. BUT Stan can't handle discrete parameters, so this is a version with `z` marginalized out, and this runs fine.